remove alt folder and move marginally closer to final file structure
